Struct burn_candle::Candle

pub struct Candle<F = f32, I = i64>where
    F: FloatCandleElement,
    I: IntCandleElement,{ /* private fields */ }

Tensor backend that uses the candle crate for executing tensor operations.

It is compatible with a wide range of hardware configurations, including CPUs and Nvidia GPUs that support CUDA. Additionally, the backend can be compiled to wasm when using the CPU.